@@ -3,7 +3,7 @@ import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
 class HiveStorage implements IPersistStore {
-  Box box;
+  late Box box;
 
   @override
   Future<void> init() async {
@@ -28,6 +28,6 @@ class HiveStorage implements IPersistStore {
 
   @override
   Future<void> deleteAll() async {
-    return box.clear();
+    box.clear();
   }
 }
