@@ -1,6 +1,7 @@
 import 'package:clean_architecture_firebase_login/data_source/fake_user_repository.dart';
 import 'package:clean_architecture_firebase_login/injected.dart';
 import 'package:clean_architecture_firebase_login/main.dart';
+import 'package:clean_architecture_firebase_login/service/exceptions/sign_in_out_exception.dart';
 import 'package:clean_architecture_firebase_login/ui/pages/sign_in_register_form_page/sign_in_register_form_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -14,20 +15,24 @@ void main() {
   });
 
   Finder emailTextFiled = find.byWidgetPredicate((widget) {
-    return widget is TextField && widget.decoration.labelText == 'Email';
+    return widget is TextField && widget.decoration?.labelText == 'Email';
   });
   Finder passwordTextFiled = find.byWidgetPredicate((widget) {
-    return widget is TextField && widget.decoration.labelText == 'Password';
+    return widget is TextField && widget.decoration?.labelText == 'Password';
+  });
+  Finder confirmPasswordTextFiled = find.byWidgetPredicate((widget) {
+    return widget is TextField &&
+        widget.decoration?.labelText == 'Confirm Password';
   });
   Finder activeSubmitButton = find.byWidgetPredicate((widget) {
-    return widget is RaisedButton && widget.enabled;
+    return widget is ElevatedButton && widget.enabled;
   });
   Finder uncheckedCheckBox = find.byWidgetPredicate((widget) {
-    return widget is Checkbox && !widget.value;
+    return widget is Checkbox && !widget.value!;
   });
 
   Finder checkedCheckBox = find.byWidgetPredicate((widget) {
-    return widget is Checkbox && widget.value;
+    return widget is Checkbox && widget.value!;
   });
 
   testWidgets('Email validation', (tester) async {
@@ -38,22 +43,22 @@ void main() {
 
     //expect that the checkBox is initially unchecked
     expect(uncheckedCheckBox, findsOneWidget);
-    //expect that the submit button is initially inactive
-    expect(activeSubmitButton, findsNothing);
+
     //expect that the submit button is shows 'Sign in' text
     expect(find.text('Sign in'), findsOneWidget);
 
     //Invalid email
     await tester.enterText(emailTextFiled, 'myemail.com');
     await tester.pump();
+    expect(find.text('Enter a valid email'), findsNothing);
+    //change focus
+    await tester.enterText(passwordTextFiled, '');
     expect(find.text('Enter a valid email'), findsOneWidget);
-    expect(activeSubmitButton, findsNothing);
 
     //valid email
     await tester.enterText(emailTextFiled, 'my@email.com');
     await tester.pump();
     expect(find.text('Enter a valid email'), findsNothing);
-    expect(activeSubmitButton, findsNothing);
   });
 
   testWidgets('password validation', (tester) async {
@@ -63,8 +68,7 @@ void main() {
     await tester.pumpAndSettle();
     //expect that the checkBox is initially unchecked
     expect(uncheckedCheckBox, findsOneWidget);
-    //expect that the submit button is initially inactive
-    expect(activeSubmitButton, findsNothing);
+
     //expect that the submit button is shows 'Sign in' text
     expect(find.text('Sign in'), findsOneWidget);
 
@@ -72,13 +76,11 @@ void main() {
     await tester.enterText(passwordTextFiled, 'mypassword');
     await tester.pump();
     expect(find.text('Enter a valid password'), findsOneWidget);
-    expect(activeSubmitButton, findsNothing);
 
     //valid password
     await tester.enterText(passwordTextFiled, 'mypassword1');
     await tester.pump();
     expect(find.text('Enter a valid password'), findsNothing);
-    expect(activeSubmitButton, findsNothing);
   });
 
   testWidgets('login with email and password', (tester) async {
@@ -93,7 +95,6 @@ void main() {
 
     expect(find.text('Enter a valid email'), findsNothing);
     expect(find.text('Enter a valid password'), findsNothing);
-    expect(activeSubmitButton, findsOneWidget);
 
     //tap on the submit button
     await tester.tap(activeSubmitButton);
@@ -130,9 +131,13 @@ void main() {
 
     expect(find.text('Enter a valid email'), findsNothing);
     expect(find.text('Enter a valid password'), findsNothing);
-    expect(activeSubmitButton, findsOneWidget);
 
     //tap on the submit button
+    await tester.tap(activeSubmitButton);
+    await tester.pump();
+    expect(find.text('Passwords do not match'), findsOneWidget);
+
+    await tester.enterText(confirmPasswordTextFiled, 'mypassword1');
     await tester.tap(activeSubmitButton);
     await tester.pump();
 
@@ -142,5 +147,31 @@ void main() {
     await tester.pumpAndSettle();
     //Expect to go to homepage after successfully signing in.
     expect(find.text('Welcome my@email.com!'), findsOneWidget);
+  });
+
+  testWidgets('server validation', (tester) async {
+    user.injectAuthMock(() =>
+        FakeUserRepository(exception: EmailException('InValid server email')));
+
+    await tester.pumpWidget(MyApp());
+    await tester.pumpAndSettle(Duration(seconds: 2));
+    await tester.tap(find.text('Sign in With Email and password'));
+    await tester.pumpAndSettle();
+    //Enter valid email and password
+    await tester.enterText(emailTextFiled, 'my@email.com');
+    await tester.enterText(passwordTextFiled, 'mypassword1');
+    await tester.pump();
+
+    expect(find.text('Enter a valid email'), findsNothing);
+    expect(find.text('Enter a valid password'), findsNothing);
+
+    //tap on the submit button
+    await tester.tap(activeSubmitButton);
+    await tester.pump();
+
+    expect(find.byType(CircularProgressIndicator), findsOneWidget);
+    //expect to find one Scaffold. This means we are still in the MyApp()
+    await tester.pumpAndSettle();
+    expect(find.text('InValid server email'), findsOneWidget);
   });
 }
