@@ -1,5 +1,5 @@
-import 'package:apple_sign_in/apple_sign_in.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+// import 'package:apple_sign_in/apple_sign_in.dart';
+import 'package:firebase_auth/firebase_auth.dart' as firebase;
 import 'package:flutter/services.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
@@ -7,23 +7,18 @@ import 'package:states_rebuilder/states_rebuilder.dart';
 import '../domain/entities/user.dart';
 import '../service/exceptions/sign_in_out_exception.dart';
 
-class UserRepository implements IAuth<User, UserParam> {
-  final FirebaseAuth _firebaseAuth = FirebaseAuth.instance;
+class UserRepository implements IAuth<User?, UserParam> {
+  final firebase.FirebaseAuth _firebaseAuth = firebase.FirebaseAuth.instance;
 
   final GoogleSignIn _googleSignIn = GoogleSignIn();
 
   @override
-  Future<IAuth<User, UserParam>> init() async {
-    return this;
-  }
-
-  @override
-  Future<User> signUp(UserParam param) {
-    switch (param.signUp) {
+  Future<User?> signUp(UserParam? param) {
+    switch (param!.signUp) {
       case SignUp.withEmailAndPassword:
         return _createUserWithEmailAndPassword(
-          param.email,
-          param.password,
+          param.email!,
+          param.password!,
         );
       default:
         throw UnimplementedError();
@@ -31,8 +26,8 @@ class UserRepository implements IAuth<User, UserParam> {
   }
 
   @override
-  Future<User> signIn(UserParam param) {
-    switch (param.signIn) {
+  Future<User?> signIn(UserParam? param) {
+    switch (param!.signIn) {
       case SignIn.anonymously:
         return _signInAnonymously();
       case SignIn.withApple:
@@ -41,8 +36,8 @@ class UserRepository implements IAuth<User, UserParam> {
         return _signInWithGoogle();
       case SignIn.withEmailAndPassword:
         return _signInWithEmailAndPassword(
-          param.email,
-          param.password,
+          param.email!,
+          param.password!,
         );
 
       default:
@@ -51,13 +46,13 @@ class UserRepository implements IAuth<User, UserParam> {
   }
 
   @override
-  Future<void> signOut(UserParam param) async {
+  Future<void> signOut(UserParam? param) async {
     final GoogleSignIn googleSignIn = GoogleSignIn();
     await googleSignIn.signOut();
     return _firebaseAuth.signOut();
   }
 
-  _fromFireBaseUserToUser(FirebaseUser user) {
+  User? _fromFireBaseUserToUser(firebase.User? user) {
     if (user == null) {
       return null;
     }
@@ -65,18 +60,19 @@ class UserRepository implements IAuth<User, UserParam> {
       uid: user.uid,
       email: user.email,
       displayName: user.displayName,
-      photoUrl: user.photoUrl,
+      photoUrl: user.photoURL,
     );
   }
 
-  Future<User> currentUser() async {
-    final FirebaseUser firebaseUser = await _firebaseAuth.currentUser();
+  Future<User?> currentUser() async {
+    final firebase.User? firebaseUser = _firebaseAuth.currentUser;
     return _fromFireBaseUserToUser(firebaseUser);
   }
 
-  Future<User> _signInAnonymously() async {
+  Future<User?> _signInAnonymously() async {
     try {
-      AuthResult authResult = await _firebaseAuth.signInAnonymously();
+      firebase.UserCredential authResult =
+          await _firebaseAuth.signInAnonymously();
       return _fromFireBaseUserToUser(authResult.user);
     } catch (e) {
       if (e is PlatformException) {
@@ -91,8 +87,8 @@ class UserRepository implements IAuth<User, UserParam> {
     }
   }
 
-  Future<User> _signInWithGoogle() async {
-    GoogleSignInAccount googleUser = await _googleSignIn.signIn();
+  Future<User?> _signInWithGoogle() async {
+    GoogleSignInAccount? googleUser = await _googleSignIn.signIn();
 
     if (googleUser == null) {
       throw SignInException(
@@ -104,13 +100,14 @@ class UserRepository implements IAuth<User, UserParam> {
     final GoogleSignInAuthentication googleAuth =
         await googleUser.authentication;
 
-    final AuthCredential credential = GoogleAuthProvider.getCredential(
+    final firebase.AuthCredential credential =
+        firebase.GoogleAuthProvider.credential(
       accessToken: googleAuth.accessToken,
       idToken: googleAuth.idToken,
     );
 
     try {
-      final AuthResult authResult =
+      final firebase.UserCredential authResult =
           (await _firebaseAuth.signInWithCredential(credential));
       return _fromFireBaseUserToUser(authResult.user);
     } catch (e) {
@@ -126,278 +123,121 @@ class UserRepository implements IAuth<User, UserParam> {
     }
   }
 
-  Future<User> _signInWithApple() async {
-    final AuthorizationResult result = await AppleSignIn.performRequests([
-      AppleIdRequest(requestedScopes: [Scope.email, Scope.fullName])
-    ]);
+  Future<User?> _signInWithApple() async {
+    // final AuthorizationResult result = await AppleSignIn.performRequests([
+    //   AppleIdRequest(requestedScopes: [Scope.email, Scope.fullName])
+    // ]);
 
-    switch (result.status) {
-      case AuthorizationStatus.authorized:
-        final appleIdCredential = result.credential;
-        final oAuthProvider = OAuthProvider(providerId: 'apple.com');
-        final credential = oAuthProvider.getCredential(
-          idToken: String.fromCharCodes(appleIdCredential.identityToken),
-          accessToken:
-              String.fromCharCodes(appleIdCredential.authorizationCode),
-        );
+    // switch (result.status) {
+    //   case AuthorizationStatus.authorized:
+    //     final appleIdCredential = result.credential;
+    //     final oAuthProvider = firebase.OAuthProvider('apple.com');
+    //     final credential = oAuthProvider.credential(
+    //       idToken: String.fromCharCodes(appleIdCredential.identityToken),
+    //       accessToken:
+    //           String.fromCharCodes(appleIdCredential.authorizationCode),
+    //     );
 
-        final authResult = await _firebaseAuth.signInWithCredential(credential);
-        final firebaseUser = authResult.user;
+    //     final authResult = await _firebaseAuth.signInWithCredential(credential);
+    //     final firebaseUser = authResult.user;
 
-        final updateUser = UserUpdateInfo();
-        updateUser.displayName =
-            '${appleIdCredential.fullName.givenName} ${appleIdCredential.fullName.familyName}';
+    //     final displayName =
+    //         '${appleIdCredential.fullName.givenName} ${appleIdCredential.fullName.familyName}';
 
-        await firebaseUser.updateProfile(updateUser);
-        return _fromFireBaseUserToUser(firebaseUser);
+    //     await firebaseUser?.updateProfile(displayName: displayName);
+    //     return _fromFireBaseUserToUser(firebaseUser);
 
-      case AuthorizationStatus.error:
-        throw SignInException(
-          title: 'Sing in with apple',
-          code: result.error.code.toString(),
-          message: result.error.localizedDescription,
-        );
-      case AuthorizationStatus.cancelled:
-        throw SignInException(
-          title: 'Sing in with apple',
-          code: '',
-          message: 'Sign in cancelled',
-        );
-    }
+    //   case AuthorizationStatus.error:
+    //     throw SignInException(
+    //       title: 'Sing in with apple',
+    //       code: result.error.code.toString(),
+    //       message: result.error.localizedDescription,
+    //     );
+    //   case AuthorizationStatus.cancelled:
+    //     throw SignInException(
+    //       title: 'Sing in with apple',
+    //       code: '',
+    //       message: 'Sign in cancelled',
+    //     );
+    // }
     return null;
   }
 
-  Future<User> _signInWithEmailAndPassword(
+  Future<User?> _signInWithEmailAndPassword(
     String email,
     String password,
   ) async {
     try {
-      final AuthResult authResult =
+      final firebase.UserCredential authResult =
           await _firebaseAuth.signInWithEmailAndPassword(
         email: email,
         password: password,
       );
       return _fromFireBaseUserToUser(authResult.user);
     } catch (e) {
-      if (e is PlatformException) {
-        throw SignInException(
-          title: 'Sign in with email and password',
-          code: e.code,
-          message: e.message,
-        );
-      } else {
-        rethrow;
+      if (e is firebase.FirebaseAuthException) {
+        switch (e.code) {
+          case 'invalid-email':
+            throw EmailException('Email address is not valid');
+          case 'user-disabled':
+            throw EmailException(
+                'User corresponding to the given email has been disabled');
+          case 'user-not-found':
+            throw EmailException(
+                'There is no user corresponding to the given email');
+          case 'wrong-password':
+            throw PasswordException('Password is invalid for the given email');
+          default:
+            throw SignInException(
+              title: 'Create use with email and password',
+              code: e.code,
+              message: e.message,
+            );
+        }
       }
+      rethrow;
     }
   }
 
-  Future<User> _createUserWithEmailAndPassword(
+  Future<User?> _createUserWithEmailAndPassword(
     String email,
     String password,
   ) async {
     try {
-      AuthResult authResult =
+      firebase.UserCredential authResult =
           await _firebaseAuth.createUserWithEmailAndPassword(
         email: email,
         password: password,
       );
       return _fromFireBaseUserToUser(authResult.user);
     } catch (e) {
-      if (e is PlatformException) {
-        throw SignInException(
-          title: 'Create use with email and password',
-          code: e.code,
-          message: e.message,
-        );
-      } else {
-        rethrow;
+      if (e is firebase.FirebaseAuthException) {
+        switch (e.code) {
+          case 'invalid-email':
+            throw EmailException('Email address is not valid');
+          case 'email-already-in-use':
+            throw EmailException(
+                'There already exists an account with the given email');
+          case 'weak-password':
+            throw EmailException(
+                'There is no user corresponding to the given email');
+          case 'wrong-password':
+            throw PasswordException('Password is not strong enough');
+          default:
+            throw SignInException(
+              title: 'Create use with email and password',
+              code: e.code,
+              message: e.message,
+            );
+        }
       }
+      rethrow;
     }
   }
 
   @override
-  void dispose() {
-    // TODO: implement dispose
-  }
+  Future<void> init() async {}
+
+  @override
+  void dispose() {}
 }
-
-// class UserRepository implements IUserRepository {
-//   final FirebaseAuth _firebaseAuth;
-
-//   final GoogleSignIn _googleSignIn;
-
-//   UserRepository({FirebaseAuth firebaseAuth, GoogleSignIn googleSignIn})
-//       : _firebaseAuth = firebaseAuth ?? FirebaseAuth.instance,
-//         _googleSignIn = googleSignIn ?? GoogleSignIn();
-
-//   @override
-//   Future<User> currentUser() async {
-//     final FirebaseUser firebaseUser = await _firebaseAuth.currentUser();
-//     return _fromFireBaseUserToUser(firebaseUser);
-//   }
-
-//   @override
-//   Future<User> signInAnonymously() async {
-//     try {
-//       AuthResult authResult = await _firebaseAuth.signInAnonymously();
-//       return _fromFireBaseUserToUser(authResult.user);
-//     } catch (e) {
-//       if (e is PlatformException) {
-//         throw SignInException(
-//           title: 'Sign in anonymously',
-//           code: e.code,
-//           message: e.message,
-//         );
-//       } else {
-//         rethrow;
-//       }
-//     }
-//   }
-
-//   @override
-//   Future<User> signInWithGoogle() async {
-//     GoogleSignInAccount googleUser = await _googleSignIn.signIn();
-
-//     if (googleUser == null) {
-//       throw SignInException(
-//         title: 'Google sign in',
-//         code: '',
-//         message: 'Sign in with google is aborted',
-//       );
-//     }
-//     final GoogleSignInAuthentication googleAuth =
-//         await googleUser.authentication;
-
-//     final AuthCredential credential = GoogleAuthProvider.getCredential(
-//       accessToken: googleAuth.accessToken,
-//       idToken: googleAuth.idToken,
-//     );
-
-//     try {
-//       final AuthResult authResult =
-//           (await _firebaseAuth.signInWithCredential(credential));
-//       return _fromFireBaseUserToUser(authResult.user);
-//     } catch (e) {
-//       if (e is PlatformException) {
-//         throw SignInException(
-//           title: 'Sign in with google',
-//           code: e.code,
-//           message: e.message,
-//         );
-//       } else {
-//         rethrow;
-//       }
-//     }
-//   }
-
-//   @override
-//   Future<User> signInWithApple() async {
-//     final AuthorizationResult result = await AppleSignIn.performRequests([
-//       AppleIdRequest(requestedScopes: [Scope.email, Scope.fullName])
-//     ]);
-
-//     switch (result.status) {
-//       case AuthorizationStatus.authorized:
-//         final appleIdCredential = result.credential;
-//         final oAuthProvider = OAuthProvider(providerId: 'apple.com');
-//         final credential = oAuthProvider.getCredential(
-//           idToken: String.fromCharCodes(appleIdCredential.identityToken),
-//           accessToken:
-//               String.fromCharCodes(appleIdCredential.authorizationCode),
-//         );
-
-//         final authResult = await _firebaseAuth.signInWithCredential(credential);
-//         final firebaseUser = authResult.user;
-
-//         final updateUser = UserUpdateInfo();
-//         updateUser.displayName =
-//             '${appleIdCredential.fullName.givenName} ${appleIdCredential.fullName.familyName}';
-
-//         await firebaseUser.updateProfile(updateUser);
-//         return _fromFireBaseUserToUser(firebaseUser);
-
-//       case AuthorizationStatus.error:
-//         throw SignInException(
-//           title: 'Sing in with apple',
-//           code: result.error.code.toString(),
-//           message: result.error.localizedDescription,
-//         );
-//       case AuthorizationStatus.cancelled:
-//         throw SignInException(
-//           title: 'Sing in with apple',
-//           code: '',
-//           message: 'Sign in cancelled',
-//         );
-//     }
-//     return null;
-//   }
-
-//   @override
-//   Future<User> createUserWithEmailAndPassword(
-//     String email,
-//     String password,
-//   ) async {
-//     try {
-//       AuthResult authResult =
-//           await _firebaseAuth.createUserWithEmailAndPassword(
-//         email: email,
-//         password: password,
-//       );
-//       return _fromFireBaseUserToUser(authResult.user);
-//     } catch (e) {
-//       if (e is PlatformException) {
-//         throw SignInException(
-//           title: 'Create use with email and password',
-//           code: e.code,
-//           message: e.message,
-//         );
-//       } else {
-//         rethrow;
-//       }
-//     }
-//   }
-
-//   @override
-//   Future<User> signInWithEmailAndPassword(
-//     String email,
-//     String password,
-//   ) async {
-//     try {
-//       final AuthResult authResult =
-//           await _firebaseAuth.signInWithEmailAndPassword(
-//         email: email,
-//         password: password,
-//       );
-//       return _fromFireBaseUserToUser(authResult.user);
-//     } catch (e) {
-//       if (e is PlatformException) {
-//         throw SignInException(
-//           title: 'Sign in with email and password',
-//           code: e.code,
-//           message: e.message,
-//         );
-//       } else {
-//         rethrow;
-//       }
-//     }
-//   }
-
-//   @override
-//   Future<void> signOut() async {
-//     final GoogleSignIn googleSignIn = GoogleSignIn();
-//     await googleSignIn.signOut();
-//     return _firebaseAuth.signOut();
-//   }
-
-//   _fromFireBaseUserToUser(FirebaseUser user) {
-//     if (user == null) {
-//       return null;
-//     }
-//     return User(
-//       uid: user.uid,
-//       email: user.email,
-//       displayName: user.displayName,
-//       photoUrl: user.photoUrl,
-//     );
-//   }
-// }
