@@ -1,3 +1,4 @@
+import 'package:clean_architecture_firebase_login/service/exceptions/sign_in_out_exception.dart';
 import 'package:flutter/material.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
@@ -10,23 +11,29 @@ import 'service/exceptions/exception_handler.dart';
 enum Env { dev, prod }
 Env currentEnv = Env.dev;
 
-final InjectedAuth<User, UserParam> user = RM.injectAuth<User, UserParam>(
+final InjectedAuth<User?, UserParam> user = RM.injectAuth<User?, UserParam>(
   () {
-    assert(currentEnv != null);
     return {
-      Env.dev: FakeUserRepository(),
-      Env.prod: UserRepository(),
-    }[currentEnv];
+      //TODO: Uncomment the exception line below to see an example of server validation
+      Env.dev: () => FakeUserRepository(
+          // exception: PasswordException('Invalid password'),
+          ),
+      Env.prod: () => UserRepository(),
+    }[currentEnv]!();
   },
-  unsignedUser: UnLoggedUser(),
   onAuthStream: (repo) => (repo as UserRepository).currentUser().asStream(),
   onSetState: On.error(
-    (err, refresh) => RM.navigate.to(
-      AlertDialog(
-        title: Text(ExceptionsHandler.errorMessage(err).title),
-        content: Text(ExceptionsHandler.errorMessage(err).message),
-      ),
-    ),
+    (err, refresh) {
+      if (err is EmailException || err is PasswordException) {
+        return;
+      }
+      RM.navigate.to(
+        AlertDialog(
+          title: Text(ExceptionsHandler.errorMessage(err).title!),
+          content: Text(ExceptionsHandler.errorMessage(err).message!),
+        ),
+      );
+    },
   ),
 );
 
@@ -34,7 +41,7 @@ final canSignInWithApple = RM.injectFuture(
   () => {
     Env.dev: FakeAppleSignInChecker().check(),
     Env.prod: AppleSignInChecker().check(),
-  }[currentEnv],
+  }[currentEnv]!,
   initialState: false,
   autoDisposeWhenNotUsed: false,
 );
