@@ -1,12 +1,39 @@
-import 'package:clean_architecture_firebase_login/domain/common/validator.dart';
-import 'package:clean_architecture_firebase_login/domain/exceptions/Validation_exception.dart';
+import 'package:clean_architecture_firebase_login/service/exceptions/sign_in_out_exception.dart';
+
+import '../../../domain/common/validator.dart';
 
 import '../../../domain/entities/user.dart';
 import 'package:flutter/material.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
 import '../../../injected.dart';
-import '../../exceptions/exceptions_handler.dart';
+
+final _email = RM.injectTextEditing(
+  validator: (String? val) {
+    if (!Validators.isValidEmail(val!)) {
+      return 'Enter a valid email';
+    }
+  },
+);
+final _password = RM.injectTextEditing(
+  validator: (String? val) {
+    if (!Validators.isValidPassword(val!)) {
+      return 'Enter a valid password';
+    }
+  },
+  validateOnTyping: true,
+);
+
+final _confirmationPassword = RM.injectTextEditing(
+  validator: (String? val) {
+    if (_password.text != val) {
+      return 'Passwords do not match';
+    }
+  },
+  validateOnTyping: true,
+);
+
+final _form = RM.injectForm();
 
 class SignInRegisterFormPage extends StatelessWidget {
   @override
@@ -21,140 +48,136 @@ class SignInRegisterFormPage extends StatelessWidget {
 }
 
 class FormWidget extends StatelessWidget {
-  //NOTE1: Creating a  ReactiveModel key for email with empty initial value
-  final _email = RM.inject<String>(
-    () => '',
-    middleSnapState: (middleSnap) {
-      //
-      if (middleSnap.nextSnap.hasData) {
-        if (!Validators.isValidEmail(middleSnap.nextSnap.data)) {
-          return middleSnap.nextSnap.copyToHasError(
-            ValidationException('Enter a valid email'),
-          );
-        }
-      }
-      return middleSnap.nextSnap;
-    },
-  );
-
-  //NOTE1: Creating a  ReactiveModel key for password with empty initial value
-  final _password = RM.inject<String>(
-    () => '',
-    middleSnapState: (middleSnap) {
-      //
-      if (middleSnap.nextSnap.hasData) {
-        if (!Validators.isValidPassword(middleSnap.nextSnap.data)) {
-          return middleSnap.nextSnap.copyToHasError(
-            ValidationException('Enter a valid password'),
-          );
-        }
-      }
-      return middleSnap.nextSnap;
-    },
-  );
-  //NOTE1: Creating a  ReactiveModel key for isRegister with false initial value
   final _isRegister = false.inj();
-  //NOTE1: bool getter to check if the form is valid
-  bool get _isFormValid => _email.hasData && _password.hasData;
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        On(
-          () => TextField(
+    return On.form(
+      () => Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          TextField(
+            controller: _email.controller,
+            focusNode: _email.focusNode,
             decoration: InputDecoration(
               icon: Icon(Icons.email),
               labelText: 'Email',
-              //NOTE4: Delegate to ExceptionsHandler.errorMessage for error handling
-              errorText: ExceptionsHandler.errorMessage(_email.error).message,
+              errorText: _email.error,
             ),
             keyboardType: TextInputType.emailAddress,
             autocorrect: false,
-            onChanged: (email) {
-              //NOTE5: set the state of email and notify observers
-              _email.state = email;
+            onSubmitted: (_) {
+              _password.focusNode.requestFocus();
             },
           ),
-        ).listenTo(_email),
-        On(
-          () => TextField(
+          TextField(
+            controller: _password.controller,
+            focusNode: _password.focusNode,
             decoration: InputDecoration(
               icon: Icon(Icons.lock),
               labelText: 'Password',
-              errorText:
-                  ExceptionsHandler.errorMessage(_password.error).message,
+              errorText: _password.error,
             ),
             obscureText: true,
             autocorrect: false,
-            onChanged: (password) {
-              _password.state = password;
+            onSubmitted: (_) {
+              if (_isRegister.state) {
+                _confirmationPassword.focusNode.requestFocus();
+              } else {
+                _form.submitFocusNode.requestFocus();
+              }
             },
           ),
-        ).listenTo(_password),
-        SizedBox(height: 10),
-        On(
-          () => Row(
-            children: <Widget>[
-              Checkbox(
-                value: _isRegister.state,
-                onChanged: (value) {
-                  _isRegister.state = value;
-                },
-              ),
-              Text(' I do not have an account')
-            ],
-          ),
-        ).listenTo(_isRegister),
-        OnCombined(
-          (_) {
-            if (user.isWaiting) {
-              return Center(child: CircularProgressIndicator());
-            }
-            return RaisedButton(
-                //NOTE8: toggle the button text between 'Register' and 'Sign in' depending on the checkbox value
-                child: _isRegister.state ? Text('Register') : Text('Sign in'),
-                //NOTE8: activate/deactivate the button if the form is valid/non valid
-                onPressed: _isFormValid
-                    ? () async {
-                        if (_isRegister.state) {
-                          await user.auth.signUp(
-                            (_) => UserParam(
-                              signUp: SignUp.withEmailAndPassword,
-                              email: _email.state,
-                              password: _password.state,
-                            ),
-                          );
-                        } else {
-                          await user.auth.signIn(
-                            (_) => UserParam(
-                              signIn: SignIn.withEmailAndPassword,
-                              email: _email.state,
-                              password: _password.state,
-                            ),
-                          );
-                        }
-                      }
-                    : null);
-          },
-        ).listenTo([_email, _password, _isRegister, user]),
-        On(
-          () {
-            // Display an error message telling the user what goes wrong.
-            if (user.hasError) {
-              return Center(
-                child: Text(
-                  ExceptionsHandler.errorMessage(user.error).message,
-                  style: TextStyle(color: Colors.red),
-                ),
-              );
-            }
-            return Text('');
-          },
-        ).listenTo(user),
-      ],
-    );
+
+          On(() => Column(
+                children: [
+                  _isRegister.state
+                      ? TextField(
+                          controller: _confirmationPassword.controller,
+                          focusNode: _confirmationPassword.focusNode,
+                          decoration: InputDecoration(
+                            icon: Icon(Icons.lock),
+                            labelText: 'Confirm Password',
+                            errorText: _confirmationPassword.error,
+                          ),
+                          obscureText: true,
+                          autocorrect: false,
+                          onSubmitted: (_) {
+                            _form.submitFocusNode.requestFocus();
+                          },
+                        )
+                      : Container(),
+                  const SizedBox(height: 10),
+                  Row(
+                    children: <Widget>[
+                      Checkbox(
+                        value: _isRegister.state,
+                        onChanged: (value) {
+                          _isRegister.state = value!;
+                        },
+                      ),
+                      Text(' I do not have an account')
+                    ],
+                  ),
+                  On.formSubmission(
+                    onSubmitting: () =>
+                        Center(child: CircularProgressIndicator()),
+                    child: ElevatedButton(
+                      focusNode: _form.submitFocusNode,
+                      child: _isRegister.state
+                          ? Text('Register')
+                          : Text('Sign in'),
+                      onPressed: () {
+                        _form.submit(
+                          () async {
+                            if (_isRegister.state) {
+                              await user.auth.signUp(
+                                (_) => UserParam(
+                                  signUp: SignUp.withEmailAndPassword,
+                                  email: _email.state,
+                                  password: _password.state,
+                                ),
+                              );
+                            } else {
+                              await user.auth.signIn(
+                                (_) => UserParam(
+                                  signIn: SignIn.withEmailAndPassword,
+                                  email: _email.state,
+                                  password: _password.state,
+                                ),
+                              );
+                              //Server validation
+                              if (user.error is EmailException) {
+                                _email.error = user.error.message;
+                              }
+                              if (user.error is PasswordException) {
+                                _password.error = user.error.message;
+                              }
+                            }
+                          },
+                        );
+                      },
+                    ),
+                  ).listenTo(_form),
+                ],
+              )).listenTo(_isRegister),
+          // On(
+          //   () {
+          //     // Display an error message telling the user what goes wrong.
+          //     if (user.hasError) {
+          //       return Center(
+          //         child: Text(
+          //           ExceptionsHandler.errorMessage(user.error).message!,
+          //           style: TextStyle(color: Colors.red),
+          //         ),
+          //       );
+          //     }
+          //     return Text('');
+          //   },
+          // ).listenTo(user),
+        ],
+      ),
+    ).listenTo(_form);
   }
 }
